Prototype a hook to handle “close” interactions on document
